+++ src/toisto/app.py
@@ -22,11 +22,10 @@ from .model.quiz.progress import Progress
 from .model.quiz.quiz import Quizzes
 from .model.quiz.quiz_factory import create_quizzes
 from .model.topic.topic import Topic
-from .persistence.concepts import ConceptLoader
 from .persistence.config import default_config, read_config
+from .persistence.loader import Loader
 from .persistence.progress import load_progress
 from .persistence.spelling_alternatives import load_spelling_alternatives
-from .persistence.topics import TopicLoader
 from .ui.cli import create_argument_parser, parse_arguments
 from .ui.text import console, show_welcome
 
@@ -35,17 +34,15 @@ def init() -> tuple[ConfigParser, Namespace, set[Concept], set[Topic], Quizzes, 
     """Initialize the main program."""
     argument_parser = create_argument_parser(default_config())
     config = read_config(argument_parser)
-    concept_loader = ConceptLoader(argument_parser)
-    concepts = concept_loader.load()
-    topic_loader = TopicLoader(argument_parser)
-    topics = topic_loader.load()
+    loader = Loader(argument_parser)
+    concepts, topics = loader.load()
     argument_parser = create_argument_parser(config, concepts, topics)
     args = parse_arguments(argument_parser)
     load_spelling_alternatives(args.target_language, args.source_language)
-    extra_concept_files = [Path(concept_file) for concept_file in args.concept_file]
-    concepts |= concept_loader.load(extra_concept_files)
-    extra_topic_files = [Path(topic_file) for topic_file in args.topic_file]
-    topics |= topic_loader.load(extra_topic_files)
+    extra_files = [Path(file_path) for file_path in args.file]
+    extra_concepts, extra_topics = loader.load(extra_files)
+    concepts |= extra_concepts
+    topics |= extra_topics
     concepts = filter_concepts(concepts, topics, args.concept, args.topic, argument_parser)
     quizzes = create_quizzes(args.target_language, args.source_language, *concepts)
     progress = load_progress(args.target_language, argument_parser)